Add "sun, test <branch>"
Summary:
I don't know how much people will use this, but it seems like a nice thing to
have, was fairly trivial to write, and might reduce the frequency of failing
tests holding up the deploy queue.  It can happen any time a make-check is not
running, including during a deploy if make-check has finished.  (Jenkins
supports queueing up behind another make-check, but I decided to not have Sun
support that.)

One issue is that make-check doesn't report back to hipchat if it succeeds.  We
might want to have an option to do that, which Sun would set.

Test Plan: deployed it and it worked

Reviewers: mroth, benjaminpollack

Reviewed By: benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D21747