Pass SLACK_CHANNEL to make-check
Summary:
I didn't think to pass this in, because it isn't an actual parameter, but it
gets passed in by deploy-via-multijob anyway.  This meant that if people used
"sun: test" their test failures would go to #1s-and-0s.  I've now added it to
jenkins as a proper parameter (rGJEe3549612), so we can pass it in from Sun.

Test Plan: try failing some tests.

Reviewers: benjaminpollack, mroth

Reviewed By: mroth

Differential Revision: https://phabricator.khanacademy.org/D21774