Link to the deployment room by id, rather than by name
Summary:
Unfortunately the Slack API we're using only supports linking to rooms by ID,
rather than by name.  (The syntax is `<#C123ABCD>`, which Slack will convert
into a link to the channel, including the name.)  So we do that.  Rather than
getting the id dynamically, I just hardcoded it the one for #1s-and-0s-deploys.

Test Plan: deploy it, and talk to sun from the wrong room.

Reviewers: benjaminpollack, mroth

Reviewed By: mroth

Differential Revision: https://phabricator.khanacademy.org/D21784