Add rudimentary queue support to Sun
Summary:
This adds support for two commands: "queue me" and "up next".  The former adds
the sender to the queue.  The latter bumps the person at the front of the queue
into the hotseat, and pings them.  If Sun can't understand the queue, it'll say
so.

I deliberately didn't hook this into the actual deploy system at all, but if we
like it, it would be fairly easy to make "finish up" automatically do an "up
next", or whatever.  The problem is more human: sometimes people want to do
something after finishing up but before the next person deploys, for instance.
So I decided to defer doing that.

Test Plan: deploy to #bot-testing, play with it.

Reviewers: mroth, benjaminpollack

Reviewed By: benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D21993