Make sun understand more complex deployer names
Summary:
Now it should accept anything of the form
`person1 + person2 + person3 (optional notes like a branch name)`,
and tag the correct people.  If it can't understand the format, it won't try to
tag anyone, but will still happily move them around in the queue.

Adding people in this format will come in a separate diff.

Test Plan:
Ran the server locally against `#bot-testing`, sent it (fake) POSTs
manually, checked that it manipulated the topic correctly.

Reviewers: benjaminpollack

Reviewed By: benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D22493