Look for set-default-start as the valid step
Summary:
`deploy_pipeline.py` now has a step `set-default-start` that gets called
instead of `set-default`.  For now we consider both of them valid steps, but in
the future we'll remove `set-default`.  So Sun should look for the new step
when deciding whether set-defualt is valid and when figuring out how to abort.

Test Plan: fingers crossed.

Reviewers: sean, benjaminpollack

Reviewed By: sean, benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D22594