Wait 20 seconds before notifying the next person in the queue that it's their turn.
Summary:
https://app.asana.com/0/210216962651575/163067552684917
We could poll jenkins to see if it's time to notify the next person or not. But simply
waiting a little while will fail in a much simplier and easier to fix way. Worse is better.

Test Plan: None. I'm don't know how to test this locally except to make sure that there are no syntax errors.

Reviewers: benjaminpollack, sean

Reviewed By: sean

Differential Revision: https://phabricator.khanacademy.org/D32246