Add "prompt znd cleanup" function to Slack
Summary:
Adding a simple Slack hook to connect to the `notify-znd-owners` Jenkins job, which will trigger ZND Zebra messages to all znd-owners to see if their znds can be cleaned up.

There'll be one follow-up change to tweak the "we've run out of znds" deploy message to suggest that users run this command.

Test Plan:
* Run sun server with `npm run monitor`
* Test "prompt znd cleanup" command by running http -f POST :8080/ team_id=T0001 team_domain=example channel_id=C2147483705 channel_name='bot-testing' user_name=sean text='sun: prompt znd cleanup' trigger_word=sun:. Ensure that Jenkins notify-znd-owners job is triggered

Reviewers: benkraft, benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D33634