Fix import error on appengine.
The json file wasn't being uploaded to appengine because of of the skip_files
setting in app.yaml.

I changed the name of the json secret file to avoid having it flagged as a file
to ignore.

Test Plan:
ran `npm run deploy` and made sure that the new instance started without error

Auditors: cslivers, benkraft, nabil