Add support for new-style deploys.
Summary:
This wasn't as bad as I thought it would be, though I had to turn some
sequential code into callbacks.

Basically, we need a new handler to start off new-style deploys, and
need to modify set-default and finish to notice what kind of deploy
we're doing and take the appropriate action.  But it's pretty easy to
tell that, so we're happy.

I refactored the jenkins infrastructure functions a bit to make it
easier to do what I wanted to do.

Fix up the validatePipelineStep promises.

First, I was returning an explicit promise where I didn't need to (and
perhaps shouldn't have).  Second, I wasn't catching the exception that
could be raised if we try to visit an `/input` page and it's not there
-- which is the expected situation when you the script isn't asking
for input at the moment.

More fixes based on actually running this in the wild.

1) We need to GET the /input/ page, not POST to it!

2) `then` -> `spread`.  This is an old bug, but nobody seems to have
   cared until now.

3) Typo for `SetDefault`.

4) Wasn't re-fetching the crumb when needed when false-clicking on the
   input button.

Oops, another typo fixing my previous typo.

Test Plan:
Ran
   env SUN_DEBUG=1 JENKINS_API_TOKEN=<value from secrets.py> npm run monitor
and I did:
   curl -d 'team_id=T0001&team_domain=example&channel_id=C090KRE5P&channel_name=1s-and-0s-deploys&user_name=csilvers&text=sun: deploy csilvers&trigger_word=sun:' localhost:8080
and it said:
   Sending: "I think there's a deploy already going on.  If that's not the case, take it up with Jenkins."
which is true.  So there's that, at least.

Ran
   env SLACK_BOT_TOKEN=... JENKINS_API_TOKEN=... DEPLOY_ROOM_ID=C096UP7D0 npm run monitor
and then started a new-style deploy and while it was deploying, I ran:
   curl -d 'team_id=T0001&team_domainxample&channel_id=C096UP7D0&channel_name=1s-and-0s-deploys&user_name=csilvers&text=sun: set default&trigger_word=sun:' localhost:8080
and saw the 1s-and-0s-deploys slack channel say:
    I'm sorry, @csilvers.  I'm afraid I can't let you do that.  (It's not time to set-default.  If you disagree, bring it up with Jenkins.)

(same)

(same)

Reviewers: benjaminpollack, benkraft

Reviewed By: benkraft

Subscribers: amos

Differential Revision: https://phabricator.khanacademy.org/D34456