Tear out all the olddeploy code.  Groovy forevah!
Summary:
Now life is much simpler!  Well, this is still a 1000 line script, but
a simpler one that it used to be.  For instance, we can get rid of
having to read the deployState object.

Test Plan:
I ran
   env SLACK_BOT_TOKEN=<K99> JENKINS_API_TOKEN=<secrets.py> DEPLOY_ROOM_ID=C096UP7D0 npm run monitor
and then ran
   curl -d 'team_id=T0001&team_domainxample&channel_id=C096UP7&channel_name=1s-and-0s-deploys&user_name=csilvers&user_id=U09M5G8G6&text=sun: state&trigger_word=sun:' localhost:8080
and saw it say a deploy was currently running (it was).

I then ran
   curl -d 'team_id=T0001&team_domainxample&channel_id=C096UP7&channel_name=1s-and-0s-deploys&user_name=csilvers&user_id=U09M5G8G6&text=sun: deploy csilverss&trigger_word=sun:' localhost:8080
and saw it say "You can't do that now," which shows that that checking
is working *and* that the IAM permissions checking is working.

I also ran
   curl -d 'team_id=T0001&team_domainxample&channel_id=C096UP7&channel_name=1s-and-0s-deploys&user_name=csilvers&user_id=U09M5G8G6&text=sun: finish&trigger_word=sun:' localhost:8080
and saw it complain that it wasn't time for that either.

Reviewers: benjaminpollack, benkraft

Reviewed By: benjaminpollack, benkraft

Differential Revision: https://phabricator.khanacademy.org/D35071