When starting a deploy, suggest to the next person that they pre-run tests.
Summary:
A lot of things had to come together to enable this!  Well, two
things.  First, we added a second bank of test machines, so we can run
two unrelated tests in parallel.  Second, we added support for merging
multiple branches together at test-time.

Fix an undefined variable problem and a wrong datatype problem.

The script wasn't working before, and it turns out it was due to these
problems: `deployBranch` wasn't defined in the scope I was accessing
it -- which I fixed by moving its definition up a level -- and
`topic.queue[0]` wasn't a string like I thought, but a more complex
data structure which I now parse properly.

In the (difficult) process of debugging all this, I fixed up some of
our error-reporting code.

Test Plan:
I'll try to deploy this right before doing a webapp deploy, so I can
debug if things go wrong.

I commented out `jenkinsJobStatus` to return `Q(null)` and
`runJobOnJenkins` to be a noop, and ran
    env SUN_DEBUG=1 SLACK_BOT_TOKEN=`cat .slack_bot_token` npm run monitor
and then
    curl -d 'team_id=T0001&team_domain=example&channel_id=C090KRE5P&channel_name=bot-testing&user_id=U09M5G8G6&text=sun: deploy foo&trigger_word=sun:' localhost:8080
It posted this text to the `#bot-testing` slack channel:
```
@undefined @benkraft ಠ_ಠ, now would be a good time to run `sun: test master + foo + <your branch>
```

Reviewers: drosile, nabil, benjaminpollack

Reviewed By: drosile, benjaminpollack

Differential Revision: https://phabricator.khanacademy.org/D39437