Don't ping the old deployer for `sun next`, or for `sun test-next`.
Summary:
A few commands that we run are actually directed toward the person
next in the queue, not the person managing the current deploy.  Those
commands send messages that explicitly at-mention the people they're
directed to.  We don't need to at-mention the people doing the current
deploy as well!  In fact, it's confusing.  So I provide a flag to turn
that off.

This is a mark that this API isn't the optimal, but it was the most
expedient solution.

Test Plan:
Mostly fingers crossed.  I did run
   env SUN_DEBUG=1 SLACK_BOT_TOKEN=`cat .slack_bot_token` npm run monitor
   curl -d 'team_id=T0001&team_domain=example&channel_id=C090KRE5P&channel_name=bot-testing&user_id=U09M5G8G6&text=sun: finish&trigger_word=sun:' localhost:8080
but it just gave an error because I didn't have the jenkins secrets loaded.

Reviewers: benjaminpollack, benkraft

Reviewed By: benkraft

Subscribers: benkraft

Differential Revision: https://phabricator.khanacademy.org/D39498