Remove babel from sun
The only thing we were using it for was ESwhatever import syntax, which
node doesn't yet do.  That was easy enough to just switch back to
require.  In addition to removing a dependency, this will make it easier
to test Sun on App Engine Standard NodeJS.

Test plan:
`SLACK_BOT_TOKEN=... JENKINS_API_TOKEN=... npm run start`, send it some
requests a la the README, it should send to `#bot-testing`

Auditors: csilvers, drosile