Retry in case of a CSRF issue, and fix up some error handling
Summary:
We cache our CSRF token forever.  For reasons we don't understand
(possibly jenkins restarts?) sometimes it is no longer valid.  In this
case, we will now refetch the CSRF token and retry.

To accomplish this, I refactored the error-handling in
`getOrPostToJenkins` a little.  This also means we return a slightly
more precise error (we distinguish connection errors from HTTP error
status codes, as well as from unhandled CSRF errors).

I also fixed up a number of places where we didn't return a promise.
Some of these were okay, but some were causing us to swallow errors,
including the Jenkins error in the above situation.  (It's really hard
to figure out which promise-returns matter!)  So now we just always
return the promise when applicable.

Test Plan:
- `SLACK_BOT_TOKEN=... JENKINS_API_TOKEN=... npm run start`, send it
  some requests a la the README, it should send to `#bot-testing`.
- cross fingers, because we don't really know how to reproduce the CSRF
  error condition.
- `SLACK_BOT_TOKEN=... JENKINS_API_TOKEN=invalidtoken npm run start`,
  send it a requests a la the README, it should send to `#bot-testing`
  to say that it can't talk to Jenkins.  (Before, it would log an error
  but say nothing to the user.)

Reviewers: csilvers, drosile

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D39827